==> analysis_code/240205-domain/01a.call_boundary.sc.ipynb <==
# _dir = '/data/female-amb/glider_25K_impute/'
# all_imputed_files = glob.glob(f'{_dir}/*/*/group*/impute/25K/group*/chunk*/*.cool')
# cell_id = [path.split('/')[-1].split('.')[0] for path in all_imputed_files]
# cell_teble = pd.DataFrame({'cell_id':cell_id,'file_paths':all_imputed_files})
# cell_teble.to_csv('All.25K.ImputedCool.csv')

import glob
import pandas as pd
import joblib
import subprocess
import pathlib

ct = 'Oligo_NN'

use_cts = joblib.load('/data/metadata/m3c_use_cts')
use_cts = [ct.replace(' ','_').replace('/','') for ct in use_cts]

meta = pd.read_csv('/data/metadata/240104_m3C_META.csv', index_col = 0)
meta['AgingMajorType'] = meta['AgingMajorType'].apply(lambda x: x.replace(' ','_').replace('/',''))
meta = meta[meta['AgingMajorType'].isin(use_cts)]
meta['AgingMajorType.Age'] = meta['AgingMajorType'] + '.' + meta['Age']

cell_table = pd.read_csv('All.25K.ImputedCool.csv', index_col = 'cell_id')[['file_paths']]
cell_table = pd.concat([cell_table, meta[['AgingMajorType','Age','AgingMajorType.Age']]], axis = 1)
cell_table = cell_table.dropna()
cell_table.head()

use_cell_table = cell_table[cell_table['AgingMajorType']== ct]

all_groups = []
for ct_age, tmpdf in use_cell_table.groupby('AgingMajorType.Age'):
    tmpdf['cell_id'] = tmpdf.index
    tmpdf = tmpdf[['cell_id','file_paths']]
    tmpdf = tmpdf.sample(300)
    tmpdf.to_csv(f"cell_table/{ct_age}.tsv", sep = '\t', header = None, index = None)
    all_groups.append(ct_age)

#output_dir = 'dataset/domain'
output_dir = 'test'
p = pathlib.Path(output_dir)
p.mkdir(parents=True, exist_ok=True)

cpu = 40

with open("boundary.sh", 'w') as file:
    for group in all_groups:
        cell_table_path = f'cool_table/{group}.tsv'
        output_prefix = f'{output_dir}/{group}'
        command = f"hicluster domain \
--cell_table_path {cell_table_path} \
--output_prefix {output_prefix} \
--resolution 25000 \
--window_size 10 \
--cpu {cpu}"
        file.write(f"{command}")   
        file.write("\n")   

#subprocess.run(command, shell = True)
